#68 - Fixed the load encounter modal name, moved the encounter view model
