+++ apps/spa-3ncount3r/src/view-models/encounter.view-model.ts
